requirements.txt
google-generativeai>=0.3.0
pandas>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
nba_api>=1.2.1
python-dotenv>=1.0.0
lxml>=4.9.0
